Some things finished
A note: the input field can no longer be clicked on; I suspect that
there's some sort of overlapping going on that I can't see.  This only
happened after I changed the CSS in order to move the 'Subscribe' button
to the right.  If anyone can fix this, I'd appreciate it.